+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # CHANGELOG
 
-##
+## Nov 30, 2020
+
+* Initial version 0.2.0 available on Github Pages